+++ src/main/java/com/kh/drugstore/inquiry/model/dao/InquiryDao.java
@@ -1,0 +1,5 @@
+package com.kh.drugstore.inquiry.model.dao;
+
+public interface InquiryDao {
+
+}
