+++ src/main/java/com/kh/drugstore/inquiry/model/dao/InquiryDaoImpl.java
@@ -1,0 +1,5 @@
+package com.kh.drugstore.inquiry.model.dao;
+
+public class InquiryDaoImpl {
+
+}
