+++ src/main/java/com/kh/drugstore/inquiry/model/dto/Answered.java
@@ -1,0 +1,6 @@
+package com.kh.drugstore.inquiry.model.dto;
+
+public enum Answered {
+
+	Y, N
+}
