+++ src/main/java/com/kh/drugstore/inquiry/model/dto/Category.java
@@ -1,4 +1,4 @@
-package com.kh.inquiry.model.dto;
+package com.kh.drugstore.inquiry.model.dto;
 
 public enum Category {
