+++ src/main/java/com/kh/drugstore/inquiry/model/dto/Inquiry.java
@@ -1,4 +1,4 @@
-package com.kh.inquiry.model.dto;
+package com.kh.drugstore.inquiry.model.dto;
 
 import java.time.LocalDateTime;
