+++ src/main/java/com/kh/drugstore/inquiry/model/service/InquiryServiceImpl.java
@@ -1,0 +1,5 @@
+package com.kh.drugstore.inquiry.model.service;
+
+public class InquiryServiceImpl {
+
+}
