--- src/main/java/com/kh/inquiry/controller/InquiryController.java
@@ -1,10 +1,0 @@
-package com.kh.inquiry.controller;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class InquiryController {
-
-	static final Logger log = LoggerFactory.getLogger(InquiryController.class);
-}
-
