--- src/main/java/com/kh/inquiry/model/dao/InquiryDao.java
@@ -1,5 +1,0 @@
-package com.kh.inquiry.model.dao;
-
-public interface InquiryDao {
-
-}
