--- src/main/java/com/kh/inquiry/model/dao/InquiryDaoImpl.java
@@ -1,5 +1,0 @@
-package com.kh.inquiry.model.dao;
-
-public class InquiryDaoImpl {
-
-}
