--- src/main/java/com/kh/inquiry/model/dto/Answered.java
@@ -1,6 +1,0 @@
-package com.kh.inquiry.model.dto;
-
-public enum Answered {
-
-	Y, N
-}
