--- src/main/java/com/kh/inquiry/model/service/InquiryService.java
@@ -1,5 +1,0 @@
-package com.kh.inquiry.model.service;
-
-public interface InquiryService {
-
-}
