--- src/main/java/com/kh/inquiry/model/service/InquiryServiceImpl.java
@@ -1,5 +1,0 @@
-package com.kh.inquiry.model.service;
-
-public class InquiryServiceImpl {
-
-}
