+++ src/main/webapp/WEB-INF/classes/mapper/inquiry/board-mapper.xml
@@ -2,6 +2,6 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="board">
+<mapper namespace="com.kh.drugstore.mapper.inquiry">
 	
 </mapper>
